API/requirements.dev.txt
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
psycopg2-binary==2.8.6
PyJWT==2.0.0
six==1.15.0
uWSGI==2.0.19.1
Werkzeug==1.0.1
